Refactor: extract helpers
